LayoutTests/http/tests/security/contentSecurityPolicy/1.1/securitypolicy-allowscriptfrom.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
  <body>
    <script src="../resources/securitypolicy-tests-base.js"></script>
    <script>
      if (document.SecurityPolicy.allowsScriptFrom('http://example.com/'))
          log('PASS script is allowed when no policy exists.');
      else
          log('FAIL script is not allowed when no policy exists.');

      injectPolicy("script-src http://notexample.com;");

      if (!document.SecurityPolicy.allowsScriptFrom('http://example.com/'))
          log('PASS script is not allowed when policy exists.');
      else
          log('FAIL script is allowed when policy exists.');
    </script>
  </body>
</html>
