safari-600.7-branch/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/securitypolicy-allowscriptfrom-expected.txt

PASS script is allowed when no policy exists.
PASS script is not allowed when policy exists.
